Fix bugprone-narrowing-conversions warnings (157 total)
Add static_cast to all sites where implicit narrowing conversions
from size_t/long long/time_t/unsigned types to int/double/float
were flagged by clang-tidy. Fixes affect 39 files across Source/,
Physics/, and Geometries/ modules.

+++ Geometries/RandomInterface/CD_RandomInterface.cpp
@@ -170,7 +170,7 @@ RandomInterface::RandomInterface() noexcept
                                                  noiseOctaves1);
 
     // Find a seed to use for the RNG
-    int seed = reseed ? std::chrono::system_clock::now().time_since_epoch().count() : 0;
+    int seed = reseed ? static_cast<int>(std::chrono::system_clock::now().time_since_epoch().count()) : 0;
 #ifdef CH_MPI
     MPI_Bcast(&seed, 1, MPI_INT, 0, Chombo_MPI::comm);
 #endif
@@ -192,7 +192,7 @@ RandomInterface::RandomInterface() noexcept
                                                  noiseOctaves2);
 
     // Find a seed to use for the RNG
-    int seed = reseed ? std::chrono::system_clock::now().time_since_epoch().count() : 1;
+    int seed = reseed ? static_cast<int>(std::chrono::system_clock::now().time_since_epoch().count()) : 1;
 #ifdef CH_MPI
     MPI_Bcast(&seed, 1, MPI_INT, 0, Chombo_MPI::comm);
 #endif
