+++ Source/AmrMesh/CD_Loads.cpp
@@ -68,7 +68,7 @@ Loads::assignLoads(const std::map<int, Real>& a_assignedLoads) noexcept
 {
   CH_TIME("Loads::assignLoads(std::map)");
 
-  const int numRanks = numProc();
+  const int numRanks = static_cast<int>(numProc());
 
   if (a_assignedLoads.size() != numRanks) {
     MayDay::Abort("Loads::assignLoads(std::map) -- a_assignedLoads.size() != numProc()");
@@ -85,7 +85,7 @@ Loads::assignLoads(const std::vector<Real>& a_assignedLoads) noexcept
 {
   CH_TIME("Loads::assignLoads(std::vector)");
 
-  const int numRanks = numProc();
+  const int numRanks = static_cast<int>(numProc());
 
   if (a_assignedLoads.size() != numRanks) {
     MayDay::Abort("Loads::assignLoads(std::vector) -- a_assignedLoads.size() != numProc()");
@@ -102,7 +102,7 @@ Loads::assignLoads(const Vector<Real>& a_assignedLoads) noexcept
 {
   CH_TIME("Loads::assignLoads(Vector)");
 
-  const int numRanks = numProc();
+  const int numRanks = static_cast<int>(numProc());
 
   if (a_assignedLoads.size() != numRanks) {
     MayDay::Abort("Loads::assignLoads(Vector) -- a_assignedLoads.size() != numProc()");
@@ -119,7 +119,7 @@ Loads::incrementLoads(const std::map<int, Real>& a_increments) noexcept
 {
   CH_TIME("Loads::incrementLoads(std::map)");
 
-  const int numRanks = numProc();
+  const int numRanks = static_cast<int>(numProc());
 
   if (a_increments.size() != numRanks) {
     MayDay::Abort("Loads::incrementLoads(std::map) -- a_increments.size() != numProc()");
@@ -136,7 +136,7 @@ Loads::incrementLoads(const std::vector<Real>& a_increments) noexcept
 {
   CH_TIME("Loads::incrementLoads(std::vector)");
 
-  const int numRanks = numProc();
+  const int numRanks = static_cast<int>(numProc());
 
   if (a_increments.size() != numRanks) {
     MayDay::Abort("Loads::incrementLoads(std::vector) -- a_increments.size() != numProc()");
@@ -153,7 +153,7 @@ Loads::incrementLoads(const Vector<Real>& a_increments) noexcept
 {
   CH_TIME("Loads::incrementLoads(Vector)");
 
-  const int numRanks = numProc();
+  const int numRanks = static_cast<int>(numProc());
 
   if (a_increments.size() != numRanks) {
     MayDay::Abort("Loads::incrementLoads(Vector) -- a_increments.size() != numProc()");
@@ -170,7 +170,7 @@ Loads::incrementLoad(const int a_rank, const Real a_increment) noexcept
 {
   CH_TIME("Loads::incrementLoad");
 
-  const int numRanks = numProc();
+  const int numRanks = static_cast<int>(numProc());
 
   if (a_rank >= numRanks) {
     MayDay::Abort("Loads::incrementLoad -- 'a_rank > numProc()'");
